+++ src/controllers/StoresController.php
@@ -26,6 +26,7 @@ use yii\web\BadRequestHttpException;
 use yii\web\NotFoundHttpException;
 use yii\web\Response;
 use yii\web\ServerErrorHttpException;
+use craft\db\Table as CraftTable;
 
 /**
  * Class Stores Controller
@@ -80,13 +81,9 @@ class StoresController extends BaseStoreManagementController
             ],
         ];
 
-        $hasOrders = $storeModel->id && (new Query())
-                ->from(['orders' => Table::ORDERS])
-                ->leftJoin(\craft\db\Table::ELEMENTS . ' el', '[[el.id]] = [[orders.id]]')
-                ->where([
-                    'storeId' => $storeModel->id,
-                    'el.dateDeleted' => null,
-                ])
+        $hasOrders = $storeModel->id && Order::find()
+                ->trashed(null)
+                ->storeId($storeModel->id)
                 ->exists();
 
         if (!$hasOrders) {